Removed resource classes and moved service providers to their own namespace Updated readme with information about the type mapper
