+++ src/Providers/ServiceProvider.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Swis\JsonApi\Client;
+namespace Swis\JsonApi\Client\Providers;
 
 use Art4\JsonApiClient\Utils\Manager as JsonApiClientManger;
 use Http\Client\HttpClient;
@@ -15,9 +15,11 @@ use Swis\JsonApi\Client\Interfaces\ClientInterface as ApiClientInterface;
 use Swis\JsonApi\Client\Interfaces\DocumentClientInterface as ApiDocumentClientInterface;
 use Swis\JsonApi\Client\Interfaces\ParserInterface;
 use Swis\JsonApi\Client\Interfaces\TypeMapperInterface;
+use Swis\JsonApi\Client\ItemDocumentSerializer;
 use Swis\JsonApi\Client\JsonApi\ErrorsParser;
 use Swis\JsonApi\Client\JsonApi\Hydrator;
 use Swis\JsonApi\Client\JsonApi\Parser;
+use Swis\JsonApi\Client\TypeMapper;
 
 class ServiceProvider extends BaseServiceProvider
 {
